requirements.txt
apiflask==2.4.0
openvino==2025.3.0
openvino-genai==2025.3.0.0
openvino-tokenizers==2025.3.0.0
langchain_community==0.3.29
optimum-intel[openvino]==1.25.2
nncf
onnx==1.18.0
modelscope==1.12.0
torch==2.4.1
transformers==4.51.3
duckduckgo-search==6.2.6
psutil>=5.9.0
numpy>=1.24.0
tqdm>=4.66.0
werkzeug>=3.0.0
pygments>=2.17.0
pytz>=2024.1
platformdirs>=4.0.0
